requirements.txt
Flask
pytesseract
opencv-python
Pillow
numpy
